task(refactor): Making project info a default task
