+++ packages/cli/src/tasks/project-meta-task.ts
@@ -8,21 +8,21 @@ import {
   getPackageJson,
 } from '@checkup/core';
 
-import ProjectInfoTaskResult from '../results/project-info-task-result';
-import { getRepositoryInfo } from '../utils/repository';
+import ProjectMetaTaskResult from '../results/project-meta-task-result';
+import { getRepositoryInfo } from '../helpers/repository';
 
-export default class ProjectInfoTask extends BaseTask implements Task {
+export default class ProjectMetaTask extends BaseTask implements Task {
   meta: TaskMetaData = {
     taskName: 'project-info',
     friendlyTaskName: 'Project Information',
     taskClassification: {
-      category: Category.Core,
+      category: Category.Meta,
       priority: Priority.High,
     },
   };
 
   async run(): Promise<TaskResult> {
-    let result: ProjectInfoTaskResult = new ProjectInfoTaskResult(this.meta);
+    let result: ProjectMetaTaskResult = new ProjectMetaTaskResult(this.meta);
     let package_ = getPackageJson(this.args.path);
 
     result.name = package_.name || '';
