+++ packages/cli/src/types/index.ts
@@ -6,3 +6,10 @@ export const enum TestType {
   Rendering = 'rendering',
   Unit = 'unit',
 }
+
+export type RepositoryInfo = {
+  totalCommits: number;
+  totalFiles: number;
+  age: string;
+  activeDays: string;
+};
