--- packages/plugin-default/README.md
@@ -1,36 +1,0 @@
-# plugin-default
-
-An oclif plugin for Ember project tasks
-
-[![oclif](https://img.shields.io/badge/cli-oclif-brightgreen.svg)](https://oclif.io)
-[![Version](https://img.shields.io/npm/v/plugin-default.svg)](https://npmjs.org/package/plugin-default)
-[![Downloads/week](https://img.shields.io/npm/dw/plugin-default.svg)](https://npmjs.org/package/plugin-default)
-[![License](https://img.shields.io/npm/l/plugin-default.svg)](https://github.com/https://github.com/checkupjs/checkup/https://github.com/checkupjs/checkup/blob/master/package.json)
-
-<!-- toc -->
-* [plugin-default](#plugin-default)
-* [Usage](#usage)
-* [Commands](#commands)
-<!-- tocstop -->
-
-# Usage
-
-<!-- usage -->
-```sh-session
-$ npm install -g @checkup/plugin-default
-$ @checkup/plugin-default COMMAND
-running command...
-$ @checkup/plugin-default (-v|--version|version)
-@checkup/plugin-default/0.0.1 darwin-x64 node-v12.16.1
-$ @checkup/plugin-default --help [COMMAND]
-USAGE
-  $ @checkup/plugin-default COMMAND
-...
-```
-<!-- usagestop -->
-
-# Commands
-
-<!-- commands -->
-
-<!-- commandsstop -->
