--- packages/plugin-default/jest.config.js
@@ -1,17 +1,0 @@
-module.exports = {
-  testEnvironment: 'node',
-  moduleFileExtensions: ['ts', 'js', 'json'],
-  transform: { '\\.ts$': 'ts-jest' },
-  coverageReporters: ['lcov', 'text-summary'],
-  collectCoverageFrom: ['src/**/*.ts'],
-  coveragePathIgnorePatterns: ['/templates/'],
-  coverageThreshold: {
-    global: {
-      branches: 100,
-      functions: 100,
-      lines: 100,
-      statements: 100,
-    },
-  },
-  testPathIgnorePatterns: ['__fixtures__'],
-};
