--- packages/plugin-default/package.json
@@ -1,60 +1,0 @@
-{
-  "name": "@checkup/plugin-default",
-  "description": "An oclif plugin for Ember project tasks",
-  "version": "0.0.1",
-  "author": "Steve Calvert @scalvert",
-  "bugs": "https://github.com/checkupjs/checkup/issues",
-  "dependencies": {
-    "@checkup/core": "0.0.1",
-    "@oclif/command": "^1",
-    "@oclif/config": "^1",
-    "promise.hash.helper": "^1.0.7",
-    "tslib": "^1"
-  },
-  "devDependencies": {
-    "@checkup/test-helpers": "0.0.1",
-    "@oclif/dev-cli": "^1",
-    "jest": "^25.2.4",
-    "rimraf": "^3.0.2",
-    "ts-jest": "^25.3.0",
-    "ts-node": "^8",
-    "type-fest": "^0.12.0",
-    "typescript": "^3.8"
-  },
-  "engines": {
-    "node": "10.* || >= 12.*"
-  },
-  "files": [
-    "/lib",
-    "/oclif.manifest.json",
-    "/yarn.lock"
-  ],
-  "homepage": "https://github.com/checkupjs/checkup",
-  "keywords": [
-    "oclif-plugin"
-  ],
-  "license": "MIT",
-  "oclif": {
-    "devPlugins": [
-      "@oclif/plugin-help"
-    ],
-    "hooks": {
-      "register-tasks": "./src/hooks/register-tasks"
-    }
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": "https://github.com/checkupjs/checkup",
-  "scripts": {
-    "build": "tsc",
-    "build:watch": "yarn build -w",
-    "clean": "rm -rf lib/*",
-    "postpack": "rm -f oclif.manifest.json",
-    "prepack": "rm -rf lib && tsc -b && oclif-dev manifest && oclif-dev readme",
-    "test": "jest",
-    "version": "oclif-dev readme && git add README.md"
-  },
-  "types": "lib/index.d.ts",
-  "main": "lib/index.js"
-}
