--- packages/plugin-default/src/hooks/register-tasks.ts
@@ -1,8 +1,0 @@
-import { Hook } from '@oclif/config';
-import { ProjectInfoTask } from '../tasks';
-
-const hook: Hook<'register-tasks'> = async function({ cliArguments, tasks }: any) {
-  tasks.registerTask(new ProjectInfoTask(cliArguments));
-};
-
-export default hook;
