--- packages/plugin-default/src/index.ts
@@ -1,3 +1,0 @@
-export default {};
-
-export * from './types';
