--- packages/plugin-default/src/tasks/index.ts
@@ -1,1 +1,0 @@
-export { default as ProjectInfoTask } from './project-info-task';
