--- packages/plugin-default/src/types/index.ts
@@ -1,6 +1,0 @@
-export type RepositoryInfo = {
-  totalCommits: number;
-  totalFiles: number;
-  age: string;
-  activeDays: string;
-};
